requirements.txt
absl-py==1.4.0
accelerate==0.19.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
boltons==23.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
crcmod==1.7
cycler==0.11.0
datasets
diffusers==0.11.1
dill
dnspython==2.3.0
docopt==0.6.2
einops==0.6.1
fastavro==1.7.4
fasteners==0.18
filelock==3.12.0
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
grpcio==1.54.2
hdfs==2.7.0
httplib2==0.21.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
iopath==0.1.10
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.2
lightning-utilities==0.8.0
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.24.3
objsize==0.6.1
omegaconf==2.3.0
open-clip-torch==2.20.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
portalocker==2.7.0
prettytable==3.6.0
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.5
pycparser==2.21
pydot==1.4.2
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
responses==0.18.0
safetensors==0.3.1
scikit-image==0.20.0
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sounddevice==0.4.6
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
tifffile==2023.4.12
timm
tokenizers==0.13.3
tomli==2.0.1
torchdiffeq
torchmetrics
torchsde
tqdm
trampoline==0.1.2
transformers==4.29.2
typing_extensions==4.5.0
tzdata
urllib3
wcwidth
xxhash==3.2.0
yacs==0.1.8
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
